Split out the test files as well.
