+++ hook_test.go
@@ -14,35 +14,6 @@ import (
 	"github.com/stvp/roll"
 )
 
-func ExampleSetupLogging() {
-	SetupLogging("some-long-token", "staging")
-
-	// This will not be reported to Rollbar
-	logrus.Info("OHAI")
-
-	// This will be reported to Rollbar
-	logrus.WithFields(logrus.Fields{"hi": "there"}).Fatal("The end.")
-}
-
-func ExampleNewHook() {
-	log := logrus.New()
-	hook := NewHook("my-secret-token", "production")
-	log.Hooks.Add(hook)
-
-	// This will not be reported to Rollbar
-	log.WithFields(logrus.Fields{"power_level": "9001"}).Debug("It's over 9000!")
-
-	// This will be reported to Rollbar
-	log.Panic("Boom.")
-}
-
-func TestLogrusHookInterface(t *testing.T) {
-	var hook interface{} = NewHook("", "foo")
-	if _, ok := hook.(logrus.Hook); !ok {
-		t.Fatal("expected NewHook's return value to implement logrus.Hook")
-	}
-}
-
 func TestIntConversion(t *testing.T) {
 	i := make(logrus.Fields)
 	i["test"] = 5
